Chore: 테스트용 데이터 SQL문 작성
* 인텔리제이 자동 들여쓰기로 들여쓰기 정리

+++ backend/src/main/resources/sql/schema.sql
@@ -3,9 +3,10 @@
 -- Model: New Model    Version: 1.0
 -- MySQL Workbench Forward Engineering
 
-SET @OLD_UNIQUE_CHECKS=@@UNIQUE_CHECKS, UNIQUE_CHECKS=0;
-SET @OLD_FOREIGN_KEY_CHECKS=@@FOREIGN_KEY_CHECKS, FOREIGN_KEY_CHECKS=0;
-SET @OLD_SQL_MODE=@@SQL_MODE, SQL_MODE='ONLY_FULL_GROUP_BY,STRICT_TRANS_TABLES,NO_ZERO_IN_DATE,NO_ZERO_DATE,ERROR_FOR_DIVISION_BY_ZERO,NO_ENGINE_SUBSTITUTION';
+SET @OLD_UNIQUE_CHECKS = @@UNIQUE_CHECKS, UNIQUE_CHECKS = 0;
+SET @OLD_FOREIGN_KEY_CHECKS = @@FOREIGN_KEY_CHECKS, FOREIGN_KEY_CHECKS = 0;
+SET @OLD_SQL_MODE = @@SQL_MODE, SQL_MODE =
+        'ONLY_FULL_GROUP_BY,STRICT_TRANS_TABLES,NO_ZERO_IN_DATE,NO_ZERO_DATE,ERROR_FOR_DIVISION_BY_ZERO,NO_ENGINE_SUBSTITUTION';
 
 -- -----------------------------------------------------
 -- Schema sidedish
@@ -14,13 +15,15 @@ SET @OLD_SQL_MODE=@@SQL_MODE, SQL_MODE='ONLY_FULL_GROUP_BY,STRICT_TRANS_TABLES,N
 -- -----------------------------------------------------
 -- Table `category`
 -- -----------------------------------------------------
-DROP TABLE IF EXISTS `category` ;
+DROP TABLE IF EXISTS `category`;
 
-CREATE TABLE IF NOT EXISTS `category` (
-                                          `id` BIGINT NOT NULL AUTO_INCREMENT,
-                                          `name` VARCHAR(45) NOT NULL,
+CREATE TABLE IF NOT EXISTS `category`
+(
+    `id`          BIGINT       NOT NULL AUTO_INCREMENT,
+    `name`        VARCHAR(45)  NOT NULL,
     `description` VARCHAR(100) NOT NULL,
-    PRIMARY KEY (`id`))
+    PRIMARY KEY (`id`)
+)
     ENGINE = InnoDB
     DEFAULT CHARACTER SET = utf8mb4
     COLLATE = utf8mb4_0900_ai_ci;
@@ -29,20 +32,22 @@ CREATE TABLE IF NOT EXISTS `category` (
 -- -----------------------------------------------------
 -- Table `subcategory`
 -- -----------------------------------------------------
-DROP TABLE IF EXISTS `subcategory` ;
+DROP TABLE IF EXISTS `subcategory`;
 
-CREATE TABLE IF NOT EXISTS `subcategory` (
-                                             `id` BIGINT NOT NULL AUTO_INCREMENT,
-                                             `name` VARCHAR(45) NOT NULL,
+CREATE TABLE IF NOT EXISTS `subcategory`
+(
+    `id`          BIGINT       NOT NULL AUTO_INCREMENT,
+    `name`        VARCHAR(45)  NOT NULL,
     `description` VARCHAR(100) NOT NULL,
-    `category_id` BIGINT NOT NULL,
+    `category_id` BIGINT       NOT NULL,
     PRIMARY KEY (`id`),
     INDEX `fk_subcategory_category1_idx` (`category_id` ASC) VISIBLE,
     CONSTRAINT `fk_subcategory_category1`
-    FOREIGN KEY (`category_id`)
-    REFERENCES `category` (`id`)
-    ON DELETE NO ACTION
-    ON UPDATE NO ACTION)
+        FOREIGN KEY (`category_id`)
+            REFERENCES `category` (`id`)
+            ON DELETE NO ACTION
+            ON UPDATE NO ACTION
+)
     ENGINE = InnoDB
     DEFAULT CHARACTER SET = utf8mb4
     COLLATE = utf8mb4_0900_ai_ci;
@@ -51,16 +56,18 @@ CREATE TABLE IF NOT EXISTS `subcategory` (
 -- -----------------------------------------------------
 -- Table `delivery_policy`
 -- -----------------------------------------------------
-DROP TABLE IF EXISTS `delivery_policy` ;
+DROP TABLE IF EXISTS `delivery_policy`;
 
-CREATE TABLE IF NOT EXISTS `delivery_policy` (
-                                                 `id` BIGINT NOT NULL AUTO_INCREMENT,
-                                                 `description` VARCHAR(100) NOT NULL,
-    `early_morning_delivery` TINYINT(1) NOT NULL,
-    `nationwide_delivery` TINYINT(1) NOT NULL,
-    `delivery_rate` INT NOT NULL,
-    `free_delivery_threshold` INT NOT NULL,
-    PRIMARY KEY (`id`))
+CREATE TABLE IF NOT EXISTS `delivery_policy`
+(
+    `id`                      BIGINT       NOT NULL AUTO_INCREMENT,
+    `description`             VARCHAR(100) NOT NULL,
+    `early_morning_delivery`  TINYINT(1)   NOT NULL,
+    `nationwide_delivery`     TINYINT(1)   NOT NULL,
+    `delivery_rate`           INT          NOT NULL,
+    `free_delivery_threshold` INT          NOT NULL,
+    PRIMARY KEY (`id`)
+)
     ENGINE = InnoDB
     DEFAULT CHARACTER SET = utf8mb4
     COLLATE = utf8mb4_0900_ai_ci;
@@ -69,38 +76,40 @@ CREATE TABLE IF NOT EXISTS `delivery_policy` (
 -- -----------------------------------------------------
 -- Table `product`
 -- -----------------------------------------------------
-DROP TABLE IF EXISTS `product` ;
+DROP TABLE IF EXISTS `product`;
 
-CREATE TABLE IF NOT EXISTS `product` (
-                                         `id` BIGINT NOT NULL AUTO_INCREMENT,
-                                         `name` VARCHAR(45) NOT NULL,
-    `description` VARCHAR(100) NOT NULL,
-    `price` INT NOT NULL,
-    `mileage_rate` DECIMAL NOT NULL,
-    `stock` INT NOT NULL,
-    `primary_image` VARCHAR(255) NOT NULL,
-    `category_id` BIGINT NOT NULL,
-    `subcategory_id` BIGINT NOT NULL,
-    `delivery_policy_id` BIGINT NOT NULL,
+CREATE TABLE IF NOT EXISTS `product`
+(
+    `id`                 BIGINT        NOT NULL AUTO_INCREMENT,
+    `name`               VARCHAR(45)   NOT NULL,
+    `description`        VARCHAR(100)  NOT NULL,
+    `price`              INT           NOT NULL,
+    `mileage_rate`       DECIMAL(5, 2) NOT NULL,
+    `stock`              INT           NOT NULL,
+    `primary_image`      VARCHAR(255)  NOT NULL,
+    `category_id`        BIGINT        NOT NULL,
+    `subcategory_id`     BIGINT        NOT NULL,
+    `delivery_policy_id` BIGINT        NOT NULL,
     PRIMARY KEY (`id`),
     INDEX `fk_product_category1_idx` (`category_id` ASC) VISIBLE,
     INDEX `fk_product_subcategory1_idx` (`subcategory_id` ASC) VISIBLE,
     INDEX `fk_product_delivery_policy1_idx` (`delivery_policy_id` ASC) VISIBLE,
     CONSTRAINT `fk_product_category1`
-    FOREIGN KEY (`category_id`)
-    REFERENCES `category` (`id`)
-    ON DELETE NO ACTION
-    ON UPDATE NO ACTION,
+        FOREIGN KEY (`category_id`)
+            REFERENCES `category` (`id`)
+            ON DELETE NO ACTION
+            ON UPDATE NO ACTION,
     CONSTRAINT `fk_product_subcategory1`
-    FOREIGN KEY (`subcategory_id`)
-    REFERENCES `subcategory` (`id`)
-    ON DELETE NO ACTION
-    ON UPDATE NO ACTION,
+        FOREIGN KEY (`subcategory_id`)
+            REFERENCES `subcategory` (`id`)
+            ON DELETE NO ACTION
+            ON UPDATE NO ACTION,
     CONSTRAINT `fk_product_delivery_policy1`
-    FOREIGN KEY (`delivery_policy_id`)
-    REFERENCES `delivery_policy` (`id`)
-    ON DELETE NO ACTION
-    ON UPDATE NO ACTION)
+        FOREIGN KEY (`delivery_policy_id`)
+            REFERENCES `delivery_policy` (`id`)
+            ON DELETE NO ACTION
+            ON UPDATE NO ACTION
+)
     ENGINE = InnoDB
     DEFAULT CHARACTER SET = utf8mb4
     COLLATE = utf8mb4_0900_ai_ci;
@@ -109,17 +118,19 @@ CREATE TABLE IF NOT EXISTS `product` (
 -- -----------------------------------------------------
 -- Table `variant_image`
 -- -----------------------------------------------------
-DROP TABLE IF EXISTS `variant_image` ;
+DROP TABLE IF EXISTS `variant_image`;
 
-CREATE TABLE IF NOT EXISTS `variant_image` (
-                                               `product_id` BIGINT NOT NULL,
-                                               `url` VARCHAR(255) NOT NULL,
+CREATE TABLE IF NOT EXISTS `variant_image`
+(
+    `product_id` BIGINT       NOT NULL,
+    `url`        VARCHAR(255) NOT NULL,
     PRIMARY KEY (`product_id`, `url`),
     CONSTRAINT `fk_image_product1`
-    FOREIGN KEY (`product_id`)
-    REFERENCES `product` (`id`)
-    ON DELETE NO ACTION
-    ON UPDATE NO ACTION)
+        FOREIGN KEY (`product_id`)
+            REFERENCES `product` (`id`)
+            ON DELETE NO ACTION
+            ON UPDATE NO ACTION
+)
     ENGINE = InnoDB
     DEFAULT CHARACTER SET = utf8mb4
     COLLATE = utf8mb4_0900_ai_ci;
@@ -128,13 +139,15 @@ CREATE TABLE IF NOT EXISTS `variant_image` (
 -- -----------------------------------------------------
 -- Table `discount`
 -- -----------------------------------------------------
-DROP TABLE IF EXISTS `discount` ;
+DROP TABLE IF EXISTS `discount`;
 
-CREATE TABLE IF NOT EXISTS `discount` (
-                                          `id` BIGINT NOT NULL AUTO_INCREMENT,
-                                          `name` VARCHAR(45) NOT NULL,
-    `discount_rate` DECIMAL NOT NULL,
-    PRIMARY KEY (`id`))
+CREATE TABLE IF NOT EXISTS `discount`
+(
+    `id`            BIGINT        NOT NULL AUTO_INCREMENT,
+    `name`          VARCHAR(45)   NOT NULL,
+    `discount_rate` DECIMAL(5, 2) NOT NULL,
+    PRIMARY KEY (`id`)
+)
     ENGINE = InnoDB
     DEFAULT CHARACTER SET = utf8mb4
     COLLATE = utf8mb4_0900_ai_ci;
@@ -143,23 +156,25 @@ CREATE TABLE IF NOT EXISTS `discount` (
 -- -----------------------------------------------------
 -- Table `product_discount`
 -- -----------------------------------------------------
-DROP TABLE IF EXISTS `product_discount` ;
+DROP TABLE IF EXISTS `product_discount`;
 
-CREATE TABLE IF NOT EXISTS `product_discount` (
-                                                  `product_id` BIGINT NOT NULL,
-                                                  `discount_id` BIGINT NOT NULL,
-                                                  PRIMARY KEY (`product_id`, `discount_id`),
+CREATE TABLE IF NOT EXISTS `product_discount`
+(
+    `product_id`  BIGINT NOT NULL,
+    `discount_id` BIGINT NOT NULL,
+    PRIMARY KEY (`product_id`, `discount_id`),
     INDEX `fk_product_discount_discount1_idx` (`discount_id` ASC) VISIBLE,
     CONSTRAINT `fk_product_discount_product1`
-    FOREIGN KEY (`product_id`)
-    REFERENCES `product` (`id`)
-    ON DELETE NO ACTION
-    ON UPDATE NO ACTION,
+        FOREIGN KEY (`product_id`)
+            REFERENCES `product` (`id`)
+            ON DELETE NO ACTION
+            ON UPDATE NO ACTION,
     CONSTRAINT `fk_product_discount_discount1`
-    FOREIGN KEY (`discount_id`)
-    REFERENCES `discount` (`id`)
-    ON DELETE NO ACTION
-    ON UPDATE NO ACTION)
+        FOREIGN KEY (`discount_id`)
+            REFERENCES `discount` (`id`)
+            ON DELETE NO ACTION
+            ON UPDATE NO ACTION
+)
     ENGINE = InnoDB
     DEFAULT CHARACTER SET = utf8mb4
     COLLATE = utf8mb4_0900_ai_ci;
@@ -168,12 +183,14 @@ CREATE TABLE IF NOT EXISTS `product_discount` (
 -- -----------------------------------------------------
 -- Table `theme`
 -- -----------------------------------------------------
-DROP TABLE IF EXISTS `theme` ;
+DROP TABLE IF EXISTS `theme`;
 
-CREATE TABLE IF NOT EXISTS `theme` (
-                                       `id` BIGINT NOT NULL AUTO_INCREMENT,
-                                       `name` VARCHAR(45) NOT NULL,
-    PRIMARY KEY (`id`))
+CREATE TABLE IF NOT EXISTS `theme`
+(
+    `id`   BIGINT      NOT NULL AUTO_INCREMENT,
+    `name` VARCHAR(45) NOT NULL,
+    PRIMARY KEY (`id`)
+)
     ENGINE = InnoDB
     DEFAULT CHARACTER SET = utf8mb4
     COLLATE = utf8mb4_0900_ai_ci;
@@ -182,28 +199,30 @@ CREATE TABLE IF NOT EXISTS `theme` (
 -- -----------------------------------------------------
 -- Table `product_theme`
 -- -----------------------------------------------------
-DROP TABLE IF EXISTS `product_theme` ;
+DROP TABLE IF EXISTS `product_theme`;
 
-CREATE TABLE IF NOT EXISTS `product_theme` (
-                                               `product_id` BIGINT NOT NULL,
-                                               `theme_id` BIGINT NOT NULL,
-                                               PRIMARY KEY (`product_id`, `theme_id`),
+CREATE TABLE IF NOT EXISTS `product_theme`
+(
+    `product_id` BIGINT NOT NULL,
+    `theme_id`   BIGINT NOT NULL,
+    PRIMARY KEY (`product_id`, `theme_id`),
     INDEX `fk_product_theme_theme1_idx` (`theme_id` ASC) VISIBLE,
     CONSTRAINT `fk_product_theme_product1`
-    FOREIGN KEY (`product_id`)
-    REFERENCES `product` (`id`)
-    ON DELETE NO ACTION
-    ON UPDATE NO ACTION,
+        FOREIGN KEY (`product_id`)
+            REFERENCES `product` (`id`)
+            ON DELETE NO ACTION
+            ON UPDATE NO ACTION,
     CONSTRAINT `fk_product_theme_theme1`
-    FOREIGN KEY (`theme_id`)
-    REFERENCES `theme` (`id`)
-    ON DELETE NO ACTION
-    ON UPDATE NO ACTION)
+        FOREIGN KEY (`theme_id`)
+            REFERENCES `theme` (`id`)
+            ON DELETE NO ACTION
+            ON UPDATE NO ACTION
+)
     ENGINE = InnoDB
     DEFAULT CHARACTER SET = utf8mb4
     COLLATE = utf8mb4_0900_ai_ci;
 
 
-SET SQL_MODE=@OLD_SQL_MODE;
-SET FOREIGN_KEY_CHECKS=@OLD_FOREIGN_KEY_CHECKS;
-SET UNIQUE_CHECKS=@OLD_UNIQUE_CHECKS;
+SET SQL_MODE = @OLD_SQL_MODE;
+SET FOREIGN_KEY_CHECKS = @OLD_FOREIGN_KEY_CHECKS;
+SET UNIQUE_CHECKS = @OLD_UNIQUE_CHECKS;
